Split Directory tests
