+++ Tests/Testably.Abstractions.Tests/FileSystem/Directory/EnumerateDirectories/DirectoryEnumerateDirectoriesTests.cs
@@ -3,9 +3,11 @@ using System.IO;
 using System.Linq;
 using Testably.Abstractions.FileSystem;
 
-namespace Testably.Abstractions.Tests.FileSystem.Directory;
+namespace Testably.Abstractions.Tests.FileSystem.Directory.EnumerateDirectories;
 
-public abstract partial class FileSystemDirectoryTests<TFileSystem>
+// ReSharper disable once PartialTypeWithSinglePart
+public abstract partial class DirectoryEnumerateDirectoriesTests<TFileSystem>
+	: FileSystemTestBase<TFileSystem>
 	where TFileSystem : IFileSystem
 {
 	[SkippableFact]
