+++ Tests/Testably.Abstractions.Tests/FileSystem/Directory/GetDirectories/DirectoryGetDirectoriesTests.cs
@@ -3,9 +3,11 @@ using System.IO;
 using System.Linq;
 using Testably.Abstractions.FileSystem;
 
-namespace Testably.Abstractions.Tests.FileSystem.Directory;
+namespace Testably.Abstractions.Tests.FileSystem.Directory.GetDirectories;
 
-public abstract partial class FileSystemDirectoryTests<TFileSystem>
+// ReSharper disable once PartialTypeWithSinglePart
+public abstract partial class DirectoryGetDirectoriesTests<TFileSystem>
+	: FileSystemTestBase<TFileSystem>
 	where TFileSystem : IFileSystem
 {
 	[SkippableTheory]
